feat: ReportActionItemCreated.js migration
* "Migrated ReportActionItemCreated.js to TS"

* "Created ReportActionItemCreated.tsx
